requirements.txt
openai==1.3.0
markdown==3.5.1
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
flask==3.0.0
